+++ src/main/kotlin/calculator/Calculator.kt
@@ -1,4 +1,4 @@
-package step2
+package calculator
 
 class Calculator {
     companion object {
