+++ src/main/kotlin/racing/model/Car.kt
@@ -1,4 +1,4 @@
-package step4.model
+package racing.model
 
 data class Car(val name: String, var raceResult: Int, val raceHistory: MutableList<Int>) {
     fun save(resultOfTurn: Int) {
