+++ src/test/kotlin/calculator/OperatorTest.kt
@@ -1,4 +1,4 @@
-package step2
+package calculator
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
