+++ src/test/kotlin/racing/model/RefereeTest.kt
@@ -1,4 +1,4 @@
-package step4.model
+package racing.model
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
